Increased defaults for DB r/w and # of threads
Increased DB read & write buffers to 1024 MB, max open files to 500 and 10 DB threads. This is to allow nodes to cope with high volumes of transactions. Helps nodes to stay synced with the network, and / or ensures node to be able to resync as fast as possible. Tested with 66k+ transactions in the pool.